+++ extract_msg/__main__.py
@@ -15,14 +15,14 @@ def main() -> None:
     # Determine where to save the files to.
     currentDir = os.getcwd() # Store this in case the path changes.
     if not args.zip:
-        if args.out_path:
-            if not os.path.exists(args.out_path):
-                os.makedirs(args.out_path)
-            out = args.out_path
+        if args.outPath:
+            if not os.path.exists(args.outPath):
+                os.makedirs(args.outPath)
+            out = args.outPath
         else:
             out = currentDir
     else:
-        out = args.out_path if args.out_path else ''
+        out = args.outPath if args.outPath else ''
 
     if args.dev:
         import extract_msg.dev
@@ -34,7 +34,7 @@ def main() -> None:
 
         from extract_msg import validation
 
-        valResults = {x[0]: validation.validate(x[0]) for x in args.msgs}
+        valResults = {x: validation.validate(x) for x in args.msgs}
         filename = f'validation {int(time.time())}.json'
         print('Validation Results:')
         pprint.pprint(valResults)
@@ -43,8 +43,8 @@ def main() -> None:
             json.dump(valResults, fil)
         input('Press enter to exit...')
     else:
-        if not args.dump_stdout:
-            utils.setupLogging(args.config_path, level, args.log, args.file_logging)
+        if not args.dumpStdout:
+            utils.setupLogging(args.configPath, level, args.log, args.fileLogging)
 
         # Quickly make a dictionary for the keyword arguments.
         kwargs = {
@@ -52,25 +52,30 @@ def main() -> None:
             'attachmentsOnly': args.attachmentsOnly,
             'charset': args.charset,
             'contentId': args.cid,
-            'customFilename': args.out_name,
+            'customFilename': args.outName,
             'customPath': out,
             'html': args.html,
             'json': args.json,
+            'pdf': args.pdf,
             'preparedHtml': args.preparedHtml,
             'rtf': args.rtf,
-            'useMsgFilename': args.use_filename,
+            'useMsgFilename': args.useFilename,
+            'wkOptions': args.wkOptions,
+            'wkPath': args.wkPath,
             'zip': args.zip,
         }
 
         for x in args.msgs:
             try:
-                with utils.openMsg(x[0]) as msg:
-                    if args.dump_stdout:
+                if args.progress:
+                    print(f'Saving file "{x}"...')
+                with utils.openMsg(x) as msg:
+                    if args.dumpStdout:
                         print(msg.body)
                     else:
                         msg.save(**kwargs)
             except Exception as e:
-                print(f'Error with file "{x[0]}": {traceback.format_exc()}')
+                print(f'Error with file "{x}": {traceback.format_exc()}')
 
 
 if __name__ == '__main__':