+++ extract_msg/message.py
@@ -2,6 +2,7 @@ import json
 import logging
 import os
 import pathlib
+import subprocess
 import zipfile
 
 import bs4
@@ -9,9 +10,9 @@ import bs4
 from imapclient.imapclient import decode_utf7
 
 from . import constants
-from .exceptions import DataNotFoundError, IncompatibleOptionsError
+from .exceptions import DataNotFoundError, IncompatibleOptionsError, WKError
 from .message_base import MessageBase
-from .utils import addNumToDir, addNumToZipDir, createZipOpen, injectHtmlHeader, injectRtfHeader, inputToBytes, inputToString, prepareFilename
+from .utils import addNumToDir, addNumToZipDir, createZipOpen, findWk, injectHtmlHeader, injectRtfHeader, inputToBytes, inputToString, prepareFilename
 
 
 logger = logging.getLogger(__name__)
@@ -109,6 +110,14 @@ class Message(MessageBase):
             usage, doing things like adding tags, injecting attachments, etc.
             This is useful for things like trying to convert the HTML body
             directly to PDF.
+        :param pdf: Used to enable saving the body as a PDF file.
+        :param wkPath: Used to manually specify the path of the wkhtmltopdf
+            executable. If not specified, the function will try to find it.
+            Useful if wkhtmltopdf is not on the path. If :param pdf: is False,
+            this argument is ignored.
+        :param wkOptions: Used to specify additional options to wkhtmltopdf.
+            this must be a list or list-like object composed of strings and
+            bytes.
         """
         # Move keyword arguments into variables.
         _json = kwargs.get('json', False)
@@ -129,6 +138,11 @@ class Message(MessageBase):
         attachOnly = kwargs.get('attachmentsOnly', False)
         # Track if we are skipping attachments.
         skipAttachments = kwargs.get('skipAttachments', False)
+
+        # If we are doing PDF, we immediately need to try to find wkhtmltopdf.
+        if pdf:
+            wkPath = findWk(kwargs.get('wkPath'))
+            kwargs['preparedHtml'] = True
 
         # ZipFile handling.
         if _zip:
@@ -252,21 +266,55 @@ class Message(MessageBase):
 
             if not attachOnly:
                 # Determine the extension to use for the body.
-                fext = 'json' if _json else fext
-
-                with _open(str(path / ('message.' + fext)), mode) as f:
-                    if _json:
-                        emailObj = json.loads(self.getJson())
-                        if not skipAttachments:
-                            emailObj['attachments'] = attachmentNames
-
-                        f.write(inputToBytes(json.dumps(emailObj), 'utf-8'))
-                    elif useHtml:
-                        f.write(self.getSaveHtmlBody(**kwargs))
-                    elif useRtf:
-                        f.write(self.getSaveRtfBody(**kwargs))
+                if pdf:
+                    # First thing is first, we need to parse our wkOptions if
+                    # they exist.
+                    wkOptions = kwargs.get('wkOptions')
+                    if wkOptions:
+                        try:
+                            # Try to convert to a list, whatever it is, and
+                            # fail if it is not possible.
+                            parsedWkOptions = [*wkOptions]
+                        except TypeError:
+                            raise TypeError(':param wkOptions: must be an iterable, not {type(wkOptions)}.')
                     else:
-                        f.write(self.getSaveBody(**kwargs))
+                        parsedWkOptions = []
+
+                    # Confirm that all of our options we now have are either
+                    # strings or bytes.
+                    if not all(isinstance(option, (str, bytes)) for option in parsedWkOptions):
+                        raise TypeError(':param wkOptions: must be an iterable of strings and bytes.')
+
+                    # We call the program to convert the html, but give tell it
+                    # the data will go in and come out through stdin and stdout,
+                    # respectively. This way we don't have to write temporary
+                    # files to the disk. We also ask that it be quiet about it.
+                    process = subprocess.Popen([wkPath, *parsedWkOptions, '-', '-'], shell = True, stdin = subprocess.PIPE, stdout = subprocess.PIPE, stderr = subprocess.PIPE)
+                    # Give the program the data and wait for the program to
+                    # finish.
+                    output = process.communicate(self.getSaveHtmlBody(**kwargs))
+                    if process.returncode != 0:
+                        raise WKError(output[1].decode('utf-8'))
+
+                    with _open(str(path / 'message.pdf'), mode) as f:
+                        f.write(output[0])
+
+                else:
+                    fext = 'json' if _json else fext
+
+                    with _open(str(path / ('message.' + fext)), mode) as f:
+                        if _json:
+                            emailObj = json.loads(self.getJson())
+                            if not skipAttachments:
+                                emailObj['attachments'] = attachmentNames
+
+                            f.write(inputToBytes(json.dumps(emailObj), 'utf-8'))
+                        elif useHtml:
+                            f.write(self.getSaveHtmlBody(**kwargs))
+                        elif useRtf:
+                            f.write(self.getSaveRtfBody(**kwargs))
+                        else:
+                            f.write(self.getSaveBody(**kwargs))
 
         except Exception:
             if not _zip: